Ignore value changes from the main switch.
Since the upgrade to a newer OpenZWave version,
I have been experiencing problems with multiple clicks
on the main switch (discovered by my 1-year-old :') ).

Apparently, the main switch will send a kind of status update when
it is pressed multiple times in succession. This is a (bad) workaround
that ignores these updates.